requirements.txt
fyers-apiv2
selenium==4.1.0
webdriver_manager==3.5.2
